refactor(core): isolate app initialization from main entrypoint
I pulled SmartRoverApp and the provider setup completely out of main.dart and moved them into a new file called app.dart.

main.dart is now strictly an entry point. It only handles the necessary hardware initialization and calls runApp(). This gets the project perfectly aligned with standard Flutter practices so the code is easier for others to navigate.
